Implemented frontend retrieval logic for loaded analysis data
Now when the Loaded Data page is opened, we'll trigger the RetrieveLoadedData endpoint and use its results to populate the tables for each state, island, and territory. The HTML has been updated to display N/A values to keep note of which fields have been populated and which haven't. Classes were added to the various table cells to identify them easier in the code.