Implemented server-retrieved Loaded Data page
This commit is a squash of three prior commits

1. Removed all manually-created evaluations
Handled via a regex replace-all with empty string for: (.*)<li>(.*)\n

2. Implemented frontend retrieval logic for loaded analysis data
Now when the Loaded Data page is opened, we'll trigger the RetrieveLoadedData endpoint and use its results to populate the tables for each state, island, and territory. The HTML has been updated to display N/A values to keep note of which fields have been populated and which haven't. Classes were added to the various table cells to identify them easier in the code.

3. Updated base address in RetrieveLoadedDataAsync()
We won't be able to point our HTTPS GitHub page to our HTTP UNLV-DEV server for requests, so we need to just point the base address at PROD and wait until 19.0 is released.